Remove "slow" ETS backend without cache
