+++ bench/list_handlers_for_event.exs
@@ -9,7 +9,11 @@ parallelism = Keyword.get(parsed, :parallelism, System.schedulers_online())
 
 segments = ?a..?z |> Enum.map(&List.wrap/1) |> Enum.map(&:erlang.list_to_atom/1)
 
-impls = [Events.Impl.Agent, Events.Impl.Ets, Events.Impl.EtsPerf, Events.Impl.EtsPerfCached]
+impls = [
+  Events.Impl.Agent,
+  Events.Impl.Ets,
+  Events.Impl.EtsPerfCached
+]
 
 Supervisor.start_link(impls, strategy: :one_for_one)
 
@@ -25,7 +29,6 @@ Benchee.run(
   %{
     "agent" => fn -> Events.Impl.Agent.list_handlers_for_event(segments) end,
     "ets" => fn -> Events.Impl.Ets.list_handlers_for_event(segments) end,
-    "ets perf" => fn -> Events.Impl.EtsPerf.list_handlers_for_event(segments) end,
     "ets perf cached" => fn -> Events.Impl.EtsPerfCached.list_handlers_for_event(segments) end
   },
   time: 10,
