+++ lib/events/impl/ets.ex
@@ -1,11 +1,23 @@
 defmodule Events.Impl.Ets do
   @moduledoc false
-  # Implementation based on a single ETS table
-  # All writes go through an Agent to guarantee uniqueness of subscription IDs.  Reads are executed directly on the ETS table by the calling process. Processes do not lock the
-  # table, so it might happend that other process attaches a handler while the process iterates
-  # through the table to find matching handlers.
+  # Implementation based on a single ETS bag table with read concurrency.
+  #
+  # Each handler is stored in a table. A key is a prefix the handler is attached to. All writes
+  # to a table go through a single Agent process to make sure that we don't get duplicate handler IDs.
+  #
+  # Reads (`list_handlers_...`) are executed by the calling process. When looking up handlers for
+  # event, first all the prefixes of event are built. Having those prefixes, a match spec is
+  # constructed to search for all handlers subscribed to these prefixes.
+  # Surprisingly (or maybe not?) ETS is so fast, that the bottleneck of `list_handlers_for_event/1`
+  # is the procedure of building all possible prefixes.
+  #
+  # When it comes to concurrency guarantess, there are basically none. When one process performs
+  # reads handlers for event, it might happen that another process deletes a handler from the table
+  # (because `:ets.select/2` is used which is not atomic).
 
   @behaviour Events.Impl
+
+  @compile {:inline, match_spec_segment: 1}
 
   @table __MODULE__
 
@@ -68,22 +80,18 @@ defmodule Events.Impl.Ets do
     end
   end
 
-  @spec match_spec_for_event(Events.event_name()) :: :ets.match_spec()
-  defp match_spec_for_event([]) do
-    [match_spec_segment([])]
+  @spec match_spec_for_event(Events.event_name(), Events.event_prefix(), [tuple()]) ::
+          :ets.match_spec()
+  defp match_spec_for_event(event, last_prefix \\ [], acc \\ [])
+
+  defp match_spec_for_event([], prev_rev_prefix, acc) do
+    [match_spec_segment(:lists.reverse(prev_rev_prefix)) | acc]
   end
 
-  defp match_spec_for_event([_] = event_name) do
-    [match_spec_segment([]), match_spec_segment(event_name)]
-  end
-
-  defp match_spec_for_event(event_name) do
-    prefixes =
-      for prefix_len <- 1..(length(event_name) - 1) do
-        match_spec_segment(Enum.take(event_name, prefix_len))
-      end
-
-    [match_spec_segment([]), match_spec_segment(event_name) | prefixes]
+  defp match_spec_for_event([segment | rest], prev_rev_prefix, acc) do
+    match_spec_for_event(rest, [segment | prev_rev_prefix], [
+      match_spec_segment(:lists.reverse(prev_rev_prefix)) | acc
+    ])
   end
 
   @spec match_spec_segment(Events.event_prefix()) :: tuple()
